luigi-nlp-preprocess: added a second luigi Task for demonstration purposes
